# openapi-snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HLOOL Mail API",
    "version": "0.9.2",
    "description": "由内部 apispec 注册表生成的机器可读 API 元数据。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "已配置的公开基础 URL"
    }
  ],
...
  "paths": {
...
    "/api/domains/available": {
      "get": {
        "tags": [
          "API Key 自动化"
        ],
        "summary": "可用域名",
        "description": "列出可选择的公共域名以及 API Key 拥有者可访问的私有域名。客户端和 AI 助手应把该端点作为生成邮箱前的域名选择入口，优先使用 data.public_domains 与 data.private_domains；data.domains 仅为旧客户端公共域名兼容字段。",
        "operationId": "availableDomains",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeAvailableDomains"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "API Key 请求同时受来源 IP、单实例入口和业务规则约束，管理员可在“API 接入”调整设置。业务额度按 API Key、HTTP 方法和路由模板计数；查询参数和具体资源 ID 不会产生独立额度。多实例共享配置但各自计数。速率超限返回 HTTP 429、rate limit exceeded 和 Retry-After 响应头；Retry-After 是至少等待的整数秒数，此类拒绝不扣每日或累计调用配额。每日或累计配额耗尽也返回 429，但不带此重试提示，需要等待配额重置或由管理员调整配额。客户端应遵守 Retry-After，使用带抖动且有次数上限的退避，避免立即并发重试；持续收件通知优先使用 Webhook。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "仅速率拒绝时返回，表示至少等待的整数秒数；配额耗尽时不返回。",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "X-API-Key"
      }
    },
...
    "/api/generate-email": {
      "post": {
        "tags": [
          "API Key 自动化"
        ],
        "summary": "生成邮箱",
        "description": "为 API Key 调用方创建或复用邮箱。默认生成普通邮箱 prefix@domain；设置 address_type=subdomain 可生成泛子域名邮箱 prefix@subdomain.domain，subdomain 留空时自动生成。传入 domain=*.example.com 会按父域 example.com 的泛子域名邮箱处理。设置 share=true 或 share.enabled=true 可在同一响应中创建邮箱分享；返回的 data.share.url 是分享页 URL，data.share.access_url 会在 URL fragment 中带上 #key=。",
        "operationId": "generateEmail",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "可选的邮箱生成参数。",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateEmailRequest"
              },
              "example": {
                "address_type": "root",
                "domain": "example.com",
                "prefix": "verify",
                "share": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "复用已有邮箱",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeGenerateEmail"
                }
              }
            }
          },
          "201": {
            "description": "邮箱已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeGenerateEmail"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "API Key 请求同时受来源 IP、单实例入口和业务规则约束，管理员可在“API 接入”调整设置。业务额度按 API Key、HTTP 方法和路由模板计数；查询参数和具体资源 ID 不会产生独立额度。多实例共享配置但各自计数。速率超限返回 HTTP 429、rate limit exceeded 和 Retry-After 响应头；Retry-After 是至少等待的整数秒数，此类拒绝不扣每日或累计调用配额。每日或累计配额耗尽也返回 429，但不带此重试提示，需要等待配额重置或由管理员调整配额。客户端应遵守 Retry-After，使用带抖动且有次数上限的退避，避免立即并发重试；持续收件通知优先使用 Webhook。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "仅速率拒绝时返回，表示至少等待的整数秒数；配额耗尽时不返回。",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "X-API-Key"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailableDomain": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "domain": {
            "type": "string"
          },
          "id": {
            "format": "uint64",
            "type": "integer"
          },
          "message_count": {
            "format": "int64",
            "type": "integer"
          },
          "mode": {
            "type": "string"
          },
          "root_ready": {
            "type": "boolean"
          },
          "wildcard_pattern": {
            "type": "string"
          },
          "wildcard_ready": {
            "type": "boolean"
          },
          "wildcard_requested": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "domain",
          "mode",
          "message_count"
        ],
        "type": "object"
      },
      "AvailableDomainsData": {
        "additionalProperties": false,
        "properties": {
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "private_domains": {
            "items": {
              "$ref": "#/components/schemas/AvailableDomain"
            },
            "type": "array"
          },
          "public_domains": {
            "items": {
              "$ref": "#/components/schemas/AvailableDomain"
            },
            "type": "array"
          },
          "public_unavailable_reason": {
            "type": "string"
          }
        },
        "required": [
          "domains",
          "public_domains",
          "private_domains"
        ],
        "type": "object"
      },
...
      "Domain": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "domain_expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uint64",
            "type": "integer"
          },
          "mailbox_created_count": {
            "format": "int64",
            "type": "integer"
          },
          "message_count": {
            "format": "int64",
            "type": "integer"
          },
          "mode": {
            "type": "string"
          },
          "mx_verified": {
            "type": "boolean"
          },
          "owner_id": {
            "format": "uint64",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "wildcard_enabled": {
            "type": "boolean"
          },
          "wildcard_requested": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "domain",
          "mode",
          "active",
          "mx_verified"
        ],
        "type": "object"
      },
      "EnvelopeAvailableDomains": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AvailableDomainsData"
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
...
      "EnvelopeGenerateEmail": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GenerateEmailData"
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "nullable": true
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
      "GenerateEmailData": {
        "additionalProperties": false,
        "properties": {
          "address_type": {
            "type": "string"
          },
          "domain": {
            "$ref": "#/components/schemas/Domain"
          },
          "domain_id": {
            "format": "uint64",
            "type": "integer"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "local_part": {
            "type": "string"
          },
          "reuse": {
            "type": "boolean"
          },
          "root_domain": {
            "type": "string"
          },
          "share": {
            "$ref": "#/components/schemas/GeneratedMailboxShare"
          },
          "subdomain": {
            "type": "string"
          },
          "wildcard_pattern": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "domain_id",
          "domain"
        ],
        "type": "object"
      },
      "GenerateEmailRequest": {
        "additionalProperties": false,
        "properties": {
          "address_type": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "share": {
            "oneOf": [
              {
                "type": "boolean"
              },
              {
                "$ref": "#/components/schemas/GenerateEmailShareOptions"
              }
            ]
          },
          "subdomain": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "GenerateEmailShareOptions": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "GeneratedMailboxShare": {
        "additionalProperties": false,
        "properties": {
          "access_url": {
            "format": "uri",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uint64",
            "type": "integer"
          },
          "key": {
            "type": "string"
          },
          "resource_type": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "token_prefix": {
            "type": "string"
          },
          "url": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "id",
          "resource_type",
          "token_prefix",
          "url",
          "access_url"
        ],
        "type": "object"
      },
...
      "Usage": {
        "additionalProperties": false,
        "properties": {
          "daily_limit": {
            "type": "string"
          },
          "daily_unlimited": {
            "type": "string"
          },
          "remaining_today": {
            "type": "string"
          },
          "remaining_total": {
            "type": "string"
          },
          "total_limit": {
            "type": "string"
          },
          "total_unlimited": {
            "type": "string"
          },
          "total_usage": {
            "type": "string"
          },
          "used_today": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "自动化端点使用的 API Key。"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gptmail_session",
        "description": "Web Console 会话专用管理端点使用的会话 Cookie。"
      }
    }
  },
...
}